Abstract logic for the creation of package event handlers.
Implement two default generators:
a) ReflectionPackageHandlerGenerator
   - Utilises classes found in the System.Reflection namespace to invoke methods.
b) EmittedPackageHandlerGenerator
   - Utilises classes found in the System.Reflection.Emit namespace to directly invoke handler methods where possible.
   - Returns NOP implementations of IPackageEventHandler where methods cannot be invoked.

See `src/MiNET/MiNET.Test/BenchmarkEventHandlerGenerators.cs` for simple benchmarks.

EmittedPackageHandlerGenerator requires the ability to generate MSIL on-the-fly and thus may not work on restricted or non-standard systems,
if people begin reporting they are unable to use it then I shall add a simple test to determine which IPackageEventHandlerGenerator is used.